[frontend] renamings (#16671)
